+++ metatomic-torch/src/model.cpp
@@ -74,6 +74,7 @@ static nlohmann::json model_output_to_json(const ModelOutputHolder& self) {
     result["unit"] = self.unit();
     result["per_atom"] = self.per_atom;
     result["explicit_gradients"] = self.explicit_gradients;
+    result["description"] = self.description;
 
     return result;
 }
@@ -96,6 +97,7 @@ static ModelOutput model_output_from_json(const nlohmann::json& data) {
     }
 
     auto result = torch::make_intrusive<ModelOutputHolder>();
+
     if (data.contains("quantity")) {
         if (!data["quantity"].is_string()) {
             throw std::runtime_error("'quantity' in JSON for ModelOutput must be a string");
@@ -123,6 +125,16 @@ static ModelOutput model_output_from_json(const nlohmann::json& data) {
             data["explicit_gradients"],
             "'explicit_gradients' in JSON for ModelOutput"
         );
+    }
+
+    if (data.contains("description")) {
+        if (!data["description"].is_string()) {
+            throw std::runtime_error("'description' in JSON for ModelOutput must be a string");
+        }
+        result->description = data["description"];
+    } else {
+        // backward compatibility
+        result->description = "";
     }
 
     return result;